requirements.txt
aider==0.2.6
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
configparser==7.1.0
dataclasses-json==0.6.7
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
filelock==3.15.4
flake8==7.1.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watson_machine_learning==1.0.360
ibm_watsonx_ai==1.1.6
idna==3.8
importlib_metadata==7.2.1
importlib_resources==6.4.4
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-huggingface==0.0.3
langchain-ibm==0.1.12
langchain-text-splitters==0.2.2
langsmith==0.1.104
litellm==1.44.7
lomond==0.3.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.5
networkx==3.2.1
numpy==1.26.4
openai==1.42.0
orjson==3.10.7
packaging==24.1
pandas==2.1.4
parsimonious==0.10.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
piptree==0.1.4
platformdirs==4.2.2
playwright==1.46.0
prompt_toolkit==3.0.47
protobuf==5.27.3
ptyprocess==0.7.0
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypager==3.0.1
pypandoc==1.13
pyperclip==1.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.2
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
setuptools-scm==8.1.0
singlestoredb==1.6.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlalchemy_singlestoredb==1.1.0
sqlparams==6.1.0
streamlit==1.37.1
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
wcwidth==0.2.13
wget==3.2
yarl==1.9.4
zipp==3.20.1
